+++ OUDS/Core/ComponentsUIKit/Sources/Indicators/_OUDSBadge.swift
@@ -24,13 +24,15 @@ extension OUDSUIKitBrige {
     /// **This is an experimental feature, feedback and support is appreciated**
     ///
     /// - Parameters:
+    ///    - accessibilityLabel: The accessibility label bringing meanings to the badge
     ///    - status: The status of this badge. The background color of the badge is based on this status, *neutral* by default
     ///    - size: The size of this badge, *medium* by default
-    @MainActor public static func createBadge(status: OUDSBadge.Status = .neutral,
+    @MainActor public static func createBadge(accessibilityLabel: String,
+                                              status: OUDSBadge.Status = .neutral,
                                               size: OUDSBadge.StandardSize = .medium) -> UIViewController
     {
         OL.warning("Avoid UIKit wrapper and prefer SwiftUI component instead OUDSBadge(status:size)")
-        let swiftUIBadge = OUDSBadge(status: status, size: size)
+        let swiftUIBadge = OUDSBadge(accessibilityLabel: accessibilityLabel, status: status, size: size)
         return wrap(component: swiftUIBadge)
     }
 
@@ -43,14 +45,16 @@ extension OUDSUIKitBrige {
     ///
     /// - Parameters:
     ///    - count:The number displayed in the badge.
+    ///    - accessibilityLabel: The accessibility label bringing meanings to the badge
     ///    - status: The status of this badge, default set to *neutral*
     ///    - size: The size of this badge, default set to *medium*
     @MainActor public static func createBadge(count: UInt,
+                                              accessibilityLabel: String,
                                               status: OUDSBadge.Status = .neutral,
                                               size: OUDSBadge.IllustrationSize = .medium) -> UIViewController
     {
         OL.warning("Avoid UIKit wrapper and prefer SwiftUI component instead OUDSBadge(count:status:size)")
-        let swiftUIBadge = OUDSBadge(count: count, status: status, size: size)
+        let swiftUIBadge = OUDSBadge(count: count, accessibilityLabel: accessibilityLabel, status: status, size: size)
         return wrap(component: swiftUIBadge)
     }