+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `checkbox` components to v2.4.0 (Orange-OpenSource/ouds-ios#1137)
 - Read only variant for `checkbox` and  `checkbox indeterminate` components (Orange-OpenSource/ouds-ios#1137)
 
+### Fixed
+
+- Add an accessibility label for badge "standard" and "count" types (Orange-OpenSource/ouds-ios#1060)
+
 ## [0.22.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.21.0...0.22.0) - 2025-11-28
 
 ### Added